+++ src/Api/ListRequestBuilder.php
@@ -2,87 +2,81 @@
 
 namespace ShopEngine\Nova\Api;
 
+use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Support\Collection;
 use Laravel\Nova\Http\Requests\NovaRequest;
 use ShopEngine\Nova\Models\ShopEngineModel;
+use ShopEngine\Nova\Resources\Code;
 use ShopEngine\Nova\Resources\Purchase;
 use ShopEngine\Nova\Structs\Api\ListRequestStruct;
 use ShopEngine\Nova\Structs\Api\RequestFilterStruct;
 use Illuminate\Container\Container;
-use Illuminate\Pagination\Paginator;
 use Laravel\Nova\Fields\Field;
 use Laravel\Nova\FilterDecoder;
 
 /**
  * @property array|RequestFilterStruct $filters
  */
-
 class ListRequestBuilder extends RequestBuilder
 {
-    public function __construct(
-        ShopEngineModel $model,
-        array $filters = []
-    ) {
+    public function __construct(ShopEngineModel $model, array $filters = [])
+    {
         parent::__construct($model);
         $this->filters = $filters;
     }
+
     /**
-     * Paginate the given query into a simple paginator.
+     * Paginate the given query into a length-aware paginator.
      *
-     * @param  int|null  $perPage
-     * @param  array  $columns
-     * @param  string  $pageName
-     * @param  int|null  $page
-     * @return \Illuminate\Contracts\Pagination\Paginator
+     * @param int|null $perPage
+     * @param array $columns
+     * @param string $pageName
+     * @param int|null $page
+     * @return LengthAwarePaginator
      */
-    public function simplePaginate(
-        $perPage = null,
-        $columns = ['*'],
-        $pageName = 'page',
-        $page = null
-    ) {
+    public function paginate($perPage = null, $columns = ['*'], $pageName = 'page', $page = null): LengthAwarePaginator
+    {
         $request = app(NovaRequest::class);
 
         //$perPage = request()->perPage();
-        $page = $page ?: Paginator::resolveCurrentPage($pageName);
+        $page = $page ?: LengthAwarePaginator::resolveCurrentPage($pageName);
 
         $listRequest = $this->buildFromRequest(
             $request,
             $perPage
         );
 
-        return $this->simplePaginator($this->loadItems($listRequest), $perPage, $page, [
-            'path' => Paginator::resolveCurrentPath(),
+        return $this->paginator($this->loadItems($listRequest), $this->count(), $perPage, $page, [
+            'path'     => LengthAwarePaginator::resolveCurrentPath(),
             'pageName' => $pageName,
         ]);
     }
 
     /**
-     * Create a new simple paginator instance.
+     * Create a new length-aware paginator instance.
      *
-     * @param  \Illuminate\Support\Collection  $items
-     * @param  int  $perPage
-     * @param  int  $currentPage
-     * @param  array  $options
-     * @return \Illuminate\Pagination\Paginator
+     * @param Collection $items
+     * @param int $total
+     * @param int $perPage
+     * @param int $currentPage
+     * @param array $options
+     * @return LengthAwarePaginator
      */
-    protected function simplePaginator($items, $perPage, $currentPage, $options): Paginator
+    protected function paginator($items, $total, $perPage, $currentPage, $options): LengthAwarePaginator
     {
-        return Container::getInstance()->makeWith(Paginator::class, compact(
-            'items',
-            'perPage',
-            'currentPage',
-            'options'
+        return Container::getInstance()->makeWith(LengthAwarePaginator::class, compact(
+            'items', 'total', 'perPage', 'currentPage', 'options'
         ));
     }
 
     /**
      * Execute the query statement on ShopEngine API.
      *
-     * @param \ShopEngine\Nova\Structs\Api\ListRequestStruct $listRequest
+     * @param ListRequestStruct $listRequest
      *
-     * @return \Illuminate\Database\Eloquent\Collection|\Illuminate\Database\Eloquent\Builder[]
+     * @return Collection
      */
-    public function loadItems(ListRequestStruct $listRequest): array
+    public function loadItems(ListRequestStruct $listRequest): Collection
     {
         $rawResponse = $this->getClient()->get(
             $this->getEndpoint(),
@@ -90,34 +84,31 @@ class ListRequestBuilder extends RequestBuilder
         );
 
         $modelClass = $this->getModelClass();
+
         return collect($rawResponse)->map(function ($seModel) use ($modelClass) {
             return new $modelClass($seModel);
-        })->all();
+        });
     }
-
 
     /**
      * Builds Request for ShopEngine
      *
-     * @param \Laravel\Nova\Http\Requests\NovaRequest $novaRequest
+     * @param NovaRequest $request
      * @param string $perPage
      *
-     * @return array
+     * @return ListRequestStruct
      */
-    public function buildFromRequest(
-        NovaRequest $request,
-        string $perPage = "25"
-    ): ListRequestStruct {
+    public function buildFromRequest(NovaRequest $request, string $perPage = '25'): ListRequestStruct
+    {
         /** @var \ShopEngine\Nova\Resources\ShopEngineResource $resource */
         $resource = $request->resource();
-
 
         $listRequest = new ListRequestStruct();
 
         if ($request->viaResourceId) {
             if (
-                $request->viaRelationship &&
-                $request->viaResource()
+                $request->viaRelationship
+                && $request->viaResource()
             ) {
                 $resourceModel = $request->newViaResource()::newModel();
 
@@ -149,10 +140,14 @@ class ListRequestBuilder extends RequestBuilder
             if ($resource::getFirstSearchKey()) {
                 $searchTerm = $request->get('search');
                 if ($searchTerm) {
+                    $searchTerm .= '%';
+                    if (!in_array($resource, [Code::class, Purchase::class])) {
+                        $searchTerm = '%' . $searchTerm;
+                    }
                     $listRequest->addFilter(
                         new RequestFilterStruct(
                             $resource::getFirstSearchKey(),
-                            urlencode("%{$searchTerm}%"),
+                            urlencode($searchTerm),
                             'like'
                         )
                     );
@@ -197,7 +192,7 @@ class ListRequestBuilder extends RequestBuilder
         }
 
         $indexFields = $request->newResource()->indexFields($request);
-        $properties = $indexFields->map(fn (Field $field) => $field->attribute)
+        $properties  = $indexFields->map(fn(Field $field) => $field->attribute)
             ->add($resource::$id)
             ->toArray();
         $listRequest->setProperties($properties);
@@ -212,7 +207,7 @@ class ListRequestBuilder extends RequestBuilder
      */
     public function count(): int
     {
-        $request = app(NovaRequest::class);
+        $request     = app(NovaRequest::class);
         $listRequest = $this->buildFromRequest($request);
 
         return $this->getClient()->get(