+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project does adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Add fast like search for code and purchase
+### Fixed
+- Fix pagination
+- Fix codepool code relation list
 
 ## [feature/articlebundle]
 ### Changed